% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{correct_formula}
\alias{correct_formula}
\title{Adjust a formula to an adduct}
\usage{
correct_formula(formula, adduct)
}
\arguments{
\item{formula}{forumla string}

\item{adduct}{adduct string of the form '[m+H]+'}
}
\value{
a list with the entries form and charge giving the new formula and
charge, respectively.
}
\description{
Adjust a formula to an adduct
}
