% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_mass}
\alias{get_mass}
\title{Calculate the mass/charge for proton adduct from molecular formula}
\usage{
get_mass(formula, adduct = "[M]", charge)
}
\arguments{
\item{formula}{String of molecular formula, insert space before  atomic
numbers}

\item{adduct}{Adduct one of "[M+H]+", "[M-H]-", "[M]+", "[M]-", "[M]"}

\item{charge}{Integer}
}
\value{
Mass/charge ratio
}
\description{
If adduct is not given or [M] assumes a [M+H]+ or [M-H]- adduct if charge is not 0.
This function has been copied from OrgMassSpecR and edited to allow the inclusion of 
isotopes.
}
\examples{
# [M+H]+ Mass of Carbamazepine-13C15N
get_mass("C14 13CH12 15NNO", 1)
}
