% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialization.R
\name{initiate}
\alias{initiate}
\title{Initiate processing of mzXML files}
\usage{
initiate(path)
}
\arguments{
\item{path}{Path to shewhart4hrms directory}
}
\value{
TRUE if it runs to the end
}
\description{
Initiate processing of mzXML files
}
\examples{
shewhart4hrms::initiate("~/linsensuppe")
}
