% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{loadReport}
\alias{loadReport}
\title{Load a report file using a dialog box or by giving the path to the file}
\usage{
loadReport(dialog = TRUE, path = NULL)
}
\description{
Load a report file using a dialog box or by giving the path to the file
}
