% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialization.R
\name{newDirTree}
\alias{newDirTree}
\title{Create directory for shewhart4hrms}
\usage{
newDirTree(path)
}
\arguments{
\item{newDir}{Path to new directory}
}
\value{
TRUE if successful
}
\description{
Creates a new shewhart4hrms directory tree and example files
}
\examples{
shewhart4hrms::create_dir("~/linsensuppe")
}
