% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak-integration.R
\name{peakpicking_BfG_cpp}
\alias{peakpicking_BfG_cpp}
\title{Pick peaks}
\usage{
peakpicking_BfG_cpp(
  i,
  rawData,
  mz_step,
  rt_min_scan,
  rt_max_scan,
  sn,
  int_threshold,
  NoiseScans,
  peakwidth_min,
  peakwidth_max,
  precursormzTol,
  maxPeaksPerSignal
)
}
\description{
Pick peaks
}
