% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report-class.R
\docType{class}
\name{Report-class}
\alias{Report-class}
\alias{Report}
\title{A reference class for shewhart4hrms}
\value{
an object of Report class
}
\description{
A class for processing mzXML data to extract peak information for 
internal standards
}
\details{
Initialize a Report by calling Report$new() this starts a new Report with the default settings 
(see settings field, below). Settings can be changed by calling the \code{changeSettings} method.
Load files by calling the \code{addRawFiles} method and process files with the \code{process_all}
method. After processing the Report can be saved for viewing in the visuallization tool (susS app).
}
\section{Fields}{

\describe{
\item{\code{rawFiles}}{A character vector with file paths in the order in which they are to be displayed.}

\item{\code{settings}}{A list with settings for all processing steps.
The settings list contains the following fields and defaults:
\code{list(rttolm = 1, mztolu = 0.05, mztolu_fine = 0.005, pol = "pos", rt_res = 1.5, EIC_extraction = 0.05, baseline_noise_MS1 = 0.5, sn = 3, rtoffset = 0,
IS_rtoffset = 0, ISrttolm = 1, area_threshold = 1, height_threshold = 1, 
use_int_threshold = "area"}.

\code{rttolm}: retention time tolerance for the suspect search. 
\code{mztolu}: m/z tolerance for the spectrum extraction (MS2 precursor mass).
\code{mztolu_fine}: m/z tolerance for precursor mass in MS1.  
\code{chromatography} Is the chromatographic method allowed from the SDB, use DOI of paper in 
which method is described.
\code{pol}: polarity. 
\code{rt_res}: Resolution of chrom. peaks (see \code{\link{ms2_search}}) in min.
Peaks with RT difference > \code{rt_res} are considered to be from different substances. 
\code{EIC_extraction}: Extraction width to produce EIC, affects integration. 
\code{baseline_noise_MS1}: Signals under this intensity are ignored and also not
included in the saved spectra. 
\code{sn}: signal-to-noise ratio limit for peak integration. 
\code{rtoffset}: Retention time offset between samples and database. 
\code{ISrttolm}: RT tolerance for IS peak finding and integration. 
\code{rtTolReinteg}: is the retention time tolerance for reintegration.
\code{mzTolReinteg}: is the m/z tolerance for reintegration. 
\code{area_threshold}: Peak area intensity threshold. 
\code{height_threshold}: Peak height intensity threshold.
\code{use_int_threshold}: can be either "area", "height", or "none"
\code{peaksPerPeak}: 
\code{mustFindChromPeak}: default FALSE, if TRUE, peaks without area are deleted}

\item{\code{rawFilesCompl}}{\code{data.frame} with processed files and date of processing.}

\item{\code{peakList}}{\code{data.frame} with all suspect search results.}

\item{\code{currentPeakID}}{numeric to keep track of peakIDs, do not change.}

\item{\code{IS}}{\code{data.frame} with list of IS to process, this has to be imported from a csv file 
(comma sep) using this method \code{addIS()}. The csv file should have the columns name, formula,
rt, adduct. Where formula is in the form e.g. for CBZ: C14 13CH12N 15NO, and adduct is in the 
form [M+H]+ or [M]+.}

\item{\code{ISresults}}{\code{data.frame} with results of IS processings.}
}}

\section{Methods}{

\describe{
\item{\code{addIS(dialog = TRUE, fileName = NULL)}}{Include a list of internal standards in the report. Dialog indicates use of interactive file
choosing dialog. The file should be a csv with 4 columns: 'name', 'formula', 'rt', 'adduct'.}

\item{\code{addRawFiles(dialog = TRUE, file_list = NULL)}}{Add raw files, (mzML or mzXML) Dialog indicates use of interactive file choosing dialog.
Does not load files to RAM, only marks file path.}

\item{\code{addRawFilesDir(dialog = TRUE, dir_path = NULL)}}{Add all raw files in the chosen directory}

\item{\code{changeSettings(parameter, value)}}{Change any setting by name and then value. See settings field for more details.}

\item{\code{clearAndSave(dialog = TRUE, nameReport = NULL)}}{Clear data from RAM and save report as .report file in the current working directory. 
Use nameReport to give a different location and different name as in 
*.clearAndSave('D:\exampleFolder\example'). Note: The folder must exist beforehand.
To read the file again, use the function shewhart4hrms::loadReport}

\item{\code{clearData(indices = NULL)}}{Remove data from RAM to clear memory, use indices of raw files}

\item{\code{getPeak(rawLinki, comp_mzi, comp_rti, minIndi, maxIndi, width, mztoli, rttoli)}}{Internal function to integrate peaks during processing}

\item{\code{loadData(all = FALSE, indices = NULL)}}{Load files that still need to be processed into RAM for fast access, if all = TRUE then
all files are loaded regardless, if indices is an integer vector, only these samps will
be loaded}

\item{\code{loadSettings()}}{Load previously settings file from current working directory. This will fail if there is more
than one file present in the directory.}

\item{\code{moveRawFile(index, direction)}}{Move a raw file in direction 'up' or 'down'}

\item{\code{process_all()}}{Process all currently unprocessed files in the report object. Previously recorded false
positives (without specified sample) are deleted by default.}

\item{\code{remRawFiles(indices)}}{Delete files based on their indices}

\item{\code{saveSettings(path = getwd())}}{Save settings to a json file. Name of file is generated from current time.}

\item{\code{view()}}{View results using shiny.}
}}

