% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shewhart4hrms-package.R
\docType{package}
\name{shewhart4hrms}
\alias{shewhart4hrms-package}
\alias{shewhart4hrms}
\title{shewhart4hrms: A package for generating Shewhart charts for LC-HRMS.}
\description{
The shewhart4hrms package provides functions for processing mzXML data and
visualizing the results together with user defined warning limits
}
\section{Usage}{

See the Readme.md file for a description of how to use shewhart4hrms
}

\author{
\strong{Maintainer}: Kevin S. Jewell \email{Jewell@bafg.de} (Federal Institute of Hydrology (BfG) Koblenz, Germany)

Authors:
\itemize{
  \item Toni Köppe \email{Koeppe@bafg.de} (Federal Institute of Hydrology (BfG) Koblenz, Germany)
  \item Christian Dietrich \email{Christian.Dietrich@bafg.de} (Federal Institute of Hydrology (BfG) Koblenz, Germany)
  \item Arne Wick \email{Wick@bafg.de} (Federal Institute of Hydrology (BfG) Koblenz, Germany)
}

}
